# Fetch all review gates together (w/ new index).