--- a/index.yaml
+++ b/index.yaml
@@ -500,6 +500,11 @@ indexes:
   - name: feature_id
 - kind: Gate
   properties:
+  - name: gate_type
+  - name: state
+  - name: stage_id
+- kind: Gate
+  properties:
   - name: state
   - name: requested_on
     direction: desc