This should speed up the reviewer's dashboard loading time by reducing the variation in latency on the high end.

Previously, the way that the pending reviews box on the "my features" page was fetching data from the backend was to:
1. Run a feature query to get the features that should show up, and then after that,
2. Make N parallel API calls to get all the gates for each feature returned in step 1.
The performance was about 1s to 2s for the feature search, plus another 200ms to 500ms waiting for the majority of gate queries to return.  But, pretty often there was one gate query that took a whole second longer, maybe due to the request starting a new GAE instance.  It also led to a noticeable visual pop when it finally rendered.

In the new approach, the queries are both started immediately in parallel:
1. The features query performance is unchanged: it still takes 1s to 2s.
2. The /gates/pending query starts immediately and  takes 1.5s to 2.5s.
So, the overall average performance is about the same.  It's hard to tell because there is a lot of variation between trials.  However, doing the single larger request seems to have more consistent performance than the old approach, and much less of a visual pop.

In this PR:
* Define a new API endpoint to get pending gates.   It returns only the gates for the stages that have an active gate, which is about half of the total number of gates returned in the previous approach.  It requires a new index, I suppose that is so that the stage_id projection can be read from the index.
* Change the feature table to make the feature query request and the gates request in parallel.
* Add the new API endpoint to cs-client.js
* Change the logic for counting weekdays between the gate request and today.  Counting days one at a time in a loop was slow for a few long-abandoned gates.  Instead, now we keep the precise count up to 30 days and use a fast approximation for longer time ranges.